Routes + Privacy + Login Add Update
Plus added Privacy Component and Login and set routes
