+++ package.json
@@ -9,6 +9,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-icons": "^4.2.0",
+    "react-router-dom": "5",
     "react-scripts": "4.0.3",
     "web-vitals": "^1.0.1"
   },
