+++ src/components/Header.js
@@ -1,5 +1,6 @@
 import PropTypes from 'prop-types'
 import Button from './Button'
+import { Link } from 'react-router-dom';
 
 const Header = ({ title, onAdd, showAdd }) => {
 
@@ -9,9 +10,12 @@ const Header = ({ title, onAdd, showAdd }) => {
 
     
     return (
+
         <header className="header">
             <h1>{title}</h1>
             <Button color={showAdd ? 'orange' : 'green'} text={showAdd ? 'Close' : 'Add'} onClick={onAdd} />
+            <br/>
+            <Link to="login">Logout</Link>
         </header>
     )
 }
