+++ src/components/PrivacyPolicy.js
@@ -1,0 +1,12 @@
+import React from 'react'
+
+const PrivacyPolicy = () => {
+    return (
+        <div>
+            this is privacy policy
+            
+        </div>
+    )
+}
+
+export default PrivacyPolicy
